+++ bank-backend/requirements.txt
@@ -9,3 +9,5 @@ pydantic~=2.10.3
 requests~=2.32.3
 psycopg2-binary
 python-multipart
+pytz
+httpx
